# contracts/openapi/rag.yaml
openapi: 3.1.0
info:
  title: RAG Service API (Private)
  version: 1.0.0
  description: Ingestion and retrieval over vector store (Qdrant) with provenance to MinIO.
servers:
  - url: http://rag:8000
security:
  - bearerAuth: []
tags:
  - name: System
  - name: RAG
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    XSessionId: { name: X-Session-Id, in: header, required: true, schema: { type: string } }
    XRequestId: { name: X-Request-Id, in: header, required: true, schema: { type: string } }
    XCorrelationId: { name: X-Correlation-Id, in: header, required: true, schema: { type: string } }
  schemas:
    Error:
      type: object
      required: [code, message, correlation_id]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        correlation_id: { type: string }
    IngestJson:
      type: object
      properties:
        text: { type: string }
        url: { type: string, format: uri }
        metadata: { type: object, additionalProperties: true }
      example:
        text: "Qdrant collection settings and ANN parameters..."
        metadata: { project: "Agentic-U3", topic: "RAG" }
    IngestAccepted:
      type: object
      properties:
        status: { type: string, enum: [accepted] }
        doc_id: { type: string }
      example: { status: accepted, doc_id: "doc-xyz" }
    RetrieveItem:
      type: object
      properties:
        text: { type: string }
        score: { type: number }
        source_url: { type: string }
        doc_id: { type: string }
        chunk_id: { type: string }
    RetrieveResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/RetrieveItem" }
      example:
        items:
          - text: "Chunked 800/50 overlap..."
            score: 0.88
            source_url: "s3://rag/ingest.md"
            doc_id: "ing-001"
            chunk_id: "c-10"
paths:
  /v1/ingest:
    post:
      tags: [RAG]
      summary: Ingest content
      description: Accept JSON text/URL or multipart file. Embeds and upserts to Qdrant.
      parameters:
        - $ref: "#/components/parameters/XSessionId"
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/XCorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IngestJson" }
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                metadata: { type: string, description: "JSON string" }
              required: [file]
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IngestAccepted" }
        "400":
          description: Bad request
          content:
            application/json: { schema: { $ref: "#/components/schemas/Error" } }
  /v1/retrieve:
    get:
      tags: [RAG]
      summary: Retrieve nearest chunks
      parameters:
        - $ref: "#/components/parameters/XSessionId"
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/XCorrelationId"
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: top_k
          in: query
          schema: { type: integer, default: 5, minimum: 1, maximum: 20 }
        - name: filters
          in: query
          style: deepObject
          explode: true
          schema: { type: object, additionalProperties: { type: string } }
      responses:
        "200":
          description: Retrieved items
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RetrieveResponse" }
  /v1/health:
    get:
      tags: [System]
      summary: Health
      security: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                example: { status: ok }
  /v1/config:
    get:
      tags: [System]
      summary: Config (embedding model, collection)
      parameters:
        - $ref: "#/components/parameters/XSessionId"
        - $ref: "#/components/parameters/XRequestId"
        - $ref: "#/components/parameters/XCorrelationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                example:
                  embedder: "all-MiniLM-L6-v2"
                  collection: "rag_chunks"
  /v1/metrics:
    get:
      tags: [System]
      summary: Prometheus metrics
      security: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      responses:
        "200":
          description: Text exposition
          content:
            text/plain:
              schema: { type: string }
              example: "rag_ingest_duration_seconds_bucket{le=\"0.5\"} 3"